feat: add OpenRouter Anthropic shim (multi-base restructure)
Restructure the OpenRouter shim from single-base (openai_chat) to
multi-base, matching the Argo/Volcengine pattern:

- openrouter/openai_chat/ — existing OpenAI Chat compatible endpoint
  (renamed from openrouter to openrouter--openai_chat)
- openrouter/anthropic/ — new Anthropic Messages API endpoint
  (base URL: https://openrouter.ai/api, native thinking support)

OpenRouter's Anthropic endpoint is a faithful proxy of the Anthropic
Messages API, so no request/response transforms are needed.

Tested against live OpenRouter API with anthropic/claude-haiku-4-5.
