+++ src/llm_rosetta/shims/providers/openrouter/anthropic/transforms.py
@@ -1,0 +1,9 @@
+"""OpenRouter Anthropic shim transforms.
+
+OpenRouter's Anthropic-compatible endpoint (`/api`) is a faithful proxy
+of the Anthropic Messages API — no request- or response-side transforms
+are needed.
+"""
+
+to_transforms = ()
+from_transforms = ()
