+++ src/llm_rosetta/shims/providers/openrouter/openai_chat/transforms.py
